% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/channel_count.R
\name{channel_count}
\alias{channel_count}
\title{Channel count}
\usage{
channel_count(x)
}
\arguments{
\item{x}{Sample sheet df}
}
\description{
R function to check barcode levels
}
\examples{
channel_count(x)
}

