% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_barcodes.R
\name{plot_barcodes}
\alias{plot_barcodes}
\title{Plot Barcodes}
\usage{
plot_barcodes(x)
}
\arguments{
\item{x}{Sample sheet df}
}
\description{
R function to test a MiSeq sample sheet
}
\examples{
plot_barcodes(x)
}
\keyword{Sample}
\keyword{Sheet}

