% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_index.R
\name{read_index}
\alias{read_index}
\title{Read index}
\usage{
read_index(index_file)
}
\arguments{
\item{index_file}{sample index file in csv format}
}
\description{
R function read a barcode file into a data frame
}
\examples{
read_index("index_file.csv")
}
\keyword{Sample}
\keyword{Sheet}

