% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_sheet_test.R
\name{sample_sheet_test}
\alias{sample_sheet_test}
\title{Sample Sheet Test}
\usage{
sample_sheet_test(barcode_file)
}
\arguments{
\item{sample_sheet}{CSV file for MiSeq}
}
\description{
R function to test a MiSeq sample sheet
}
\examples{
sample_sheet_test(sample_sheet)
}
\keyword{Sample}
\keyword{Sheet}

