% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barcode_select.R
\name{barcode_select}
\alias{barcode_select}
\title{Barcode Select}
\usage{
barcode_select(indexes = c(1:75), number_to_select = 6)
}
\arguments{
\item{indexes}{indexes to select from defaults [1-75]}

\item{number_to_select}{number of barcodes to select}
}
\description{
R function to select random barcodes that don't overlap colour channels
}
\examples{
barcode_select(indexes = c(1:75), number_to_select = 6)
}

